Moving files from component to Reader
