+++ client/app/reader/DocumentsTable.jsx
@@ -4,9 +4,9 @@ import PropTypes from 'prop-types';
 import _ from 'lodash';
 import { connect } from 'react-redux';
 import { formatDateStr } from '../util/DateUtil';
-import Comment from '../components/Comment';
+import Comment from './Comment';
 import DocumentCategoryIcons from '../components/DocumentCategoryIcons';
-import TagTableColumn from '../components/reader/TagTableColumn';
+import TagTableColumn from './TagTableColumn';
 import Table from '../components/Table';
 import Button from '../components/Button';
 import * as Constants from './constants';
