+++ client/app/reader/PdfListView.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
-import DocumentListHeader from '../components/reader/DocumentListHeader';
+import DocumentListHeader from './DocumentListHeader';
 import ClaimsFolderDetails from './ClaimsFolderDetails';
 import { fetchAppealDetails } from './actions';
 import { shouldFetchAppeal } from '../reader/utils';
